Constructor Function for Player class

Constructor Function for Player class

player(
  id,
  first_name,
  last_name,
  team,
  position,
  depth = 1L,
  salary,
  fpts,
  locked = FALSE,
  blocked = FALSE,
  is_injured = FALSE,
  min_exposure = NA_real_,
  max_exposure = NA_real_,
  variance = NA_real_,
  game_info
)

Arguments

id
player ID

first_name
first name

last_name
last name

team
team (Required)

position
position (Required)

depth
Position on depth chart or roster order (like batting order) (Default 1)

salary
salary

fpts
fantasy points

locked
Whether the player should be locked to all lineups (Default FALSE)

blocked
Whether to omit player from all lineups (Default FALSE)

is_injured
injury flag (Default FALSE)

min_exposure
Minimum exposure across lineups

max_exposure
Maximum exposure across lineups

variance
Amount of variance to apply to fpts during optimization (percentage)

game_info
information about the game (Default is an empty game_info object)

Details

Function that initializes a Player object.